+++ src/main/java/ladysnake/satin/api/experimental/managed/Uniform3f.java
@@ -18,18 +18,15 @@
 package ladysnake.satin.api.experimental.managed;
 
 import org.apiguardian.api.API;
+import org.jetbrains.annotations.ApiStatus;
 
-import static org.apiguardian.api.API.Status.EXPERIMENTAL;
+import static org.apiguardian.api.API.Status.DEPRECATED;
 
-public interface Uniform3f {
-
-    /**
-     * Sets the value of a uniform declared in json
-     *
-     * @param value0 float value
-     * @param value1 float value
-     * @param value2 float value
-     */
-    @API(status = EXPERIMENTAL, since = "1.3.0")
-    void set(float value0, float value1, float value2);
+/**
+ * @deprecated use the promoted version in {@code ladysnake.satin.api.managed.uniform}
+ */
+@Deprecated
+@ApiStatus.ScheduledForRemoval
+@API(status = DEPRECATED, since = "1.4.0")
+public interface Uniform3f extends ladysnake.satin.api.managed.uniform.Uniform3f {
 }