+++ src/main/java/ladysnake/satin/api/managed/uniform/UniformFinder.java
@@ -15,27 +15,20 @@
  * You should have received a copy of the GNU Lesser General Public License
  * along with this program; If not, see <https://www.gnu.org/licenses>.
  */
-package ladysnake.satin.api.experimental.managed;
+package ladysnake.satin.api.managed.uniform;
 
 import org.apiguardian.api.API;
-import org.jetbrains.annotations.ApiStatus;
 
-import static org.apiguardian.api.API.Status.DEPRECATED;
 import static org.apiguardian.api.API.Status.EXPERIMENTAL;
+import static org.apiguardian.api.API.Status.MAINTAINED;
 
-/**
- * @deprecated use the promoted version in {@code ladysnake.satin.api.managed.uniform}
- */
-@Deprecated
-@ApiStatus.ScheduledForRemoval
-@API(status = DEPRECATED, since = "1.4.0")
-public interface UniformFinder extends ladysnake.satin.api.managed.uniform.UniformFinder {
+public interface UniformFinder {
     /**
      * Finds a uniform declared in json
      *
      * @param uniformName the name of the uniform field in the shader source file
      */
-    @API(status = EXPERIMENTAL, since = "1.3.0")
+    @API(status = MAINTAINED, since = "1.4.0")
     Uniform1i findUniform1i(String uniformName);
 
     /**
@@ -43,7 +36,7 @@ public interface UniformFinder extends ladysnake.satin.api.managed.uniform.Unifo
      *
      * @param uniformName the name of the uniform field in the shader source file
      */
-    @API(status = EXPERIMENTAL, since = "1.3.0")
+    @API(status = MAINTAINED, since = "1.4.0")
     Uniform2i findUniform2i(String uniformName);
 
     /**
@@ -51,7 +44,7 @@ public interface UniformFinder extends ladysnake.satin.api.managed.uniform.Unifo
      *
      * @param uniformName the name of the uniform field in the shader source file
      */
-    @API(status = EXPERIMENTAL, since = "1.3.0")
+    @API(status = MAINTAINED, since = "1.4.0")
     Uniform3i findUniform3i(String uniformName);
 
     /**
@@ -59,7 +52,7 @@ public interface UniformFinder extends ladysnake.satin.api.managed.uniform.Unifo
      *
      * @param uniformName the name of the uniform field in the shader source file
      */
-    @API(status = EXPERIMENTAL, since = "1.3.0")
+    @API(status = MAINTAINED, since = "1.4.0")
     Uniform4i findUniform4i(String uniformName);
 
     /**
@@ -67,7 +60,7 @@ public interface UniformFinder extends ladysnake.satin.api.managed.uniform.Unifo
      *
      * @param uniformName the name of the uniform field in the shader source file
      */
-    @API(status = EXPERIMENTAL, since = "1.3.0")
+    @API(status = MAINTAINED, since = "1.4.0")
     Uniform1f findUniform1f(String uniformName);
 
     /**
@@ -75,7 +68,7 @@ public interface UniformFinder extends ladysnake.satin.api.managed.uniform.Unifo
      *
      * @param uniformName the name of the uniform field in the shader source file
      */
-    @API(status = EXPERIMENTAL, since = "1.3.0")
+    @API(status = MAINTAINED, since = "1.4.0")
     Uniform2f findUniform2f(String uniformName);
 
     /**
@@ -83,7 +76,7 @@ public interface UniformFinder extends ladysnake.satin.api.managed.uniform.Unifo
      *
      * @param uniformName the name of the uniform field in the shader source file
      */
-    @API(status = EXPERIMENTAL, since = "1.3.0")
+    @API(status = MAINTAINED, since = "1.4.0")
     Uniform3f findUniform3f(String uniformName);
 
     /**
@@ -91,7 +84,7 @@ public interface UniformFinder extends ladysnake.satin.api.managed.uniform.Unifo
      *
      * @param uniformName the name of the uniform field in the shader source file
      */
-    @API(status = EXPERIMENTAL, since = "1.3.0")
+    @API(status = MAINTAINED, since = "1.4.0")
     Uniform4f findUniform4f(String uniformName);
 
     /**
@@ -99,7 +92,14 @@ public interface UniformFinder extends ladysnake.satin.api.managed.uniform.Unifo
      *
      * @param uniformName the name of the uniform field in the shader source file
      */
-    @API(status = EXPERIMENTAL, since = "1.3.0")
+    @API(status = MAINTAINED, since = "1.4.0")
     UniformMat4 findUniformMat4(String uniformName);
+
+    /**
+     * Finds a sampler declared in json
+     * @param samplerName the name of the sampler field in the shader source file
+     */
+    @API(status = EXPERIMENTAL, since = "1.4.0")
+    SamplerUniform findSampler(String samplerName);
 
 }